% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{add_prediction}
\alias{add_prediction}
\title{Add predictions from a fitted model to a data set}
\usage{
add_prediction(data, object)
}
\arguments{
\item{data}{dataset to add predictions to}

\item{object}{a model object}
}
\description{
Add predictions from a fitted model to a data set
}

