% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{co_cite}
\alias{co_cite}
\title{Add network metadata to edge and node lists}
\usage{
co_cite(edges)
}
\arguments{
\item{edges}{data frame describing the edges}
}
\description{
Add network metadata to edge and node lists
}
