% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{ConfInt}
\alias{ConfInt}
\alias{sm_mean_ci_boot}
\alias{sm_mean_ci_normal}
\alias{sm_median_hilow}
\title{Confidence interval summaries}
\usage{
sm_mean_ci_boot(x, ..., name = NULL)

sm_mean_ci_normal(x, ..., name = NULL)

sm_median_hilow(x, ..., name = NULL)
}
\arguments{
\item{x}{the variable to summarize. Will be used deparsed to infer the
names of of the summaries.}

\item{...}{other arguments passed on to the respective Hmisc function.}

\item{name}{provide a prefix variable name for the output. Set to
\code{NULL} to infer from the input.}
}
\description{
Summarize a variable by giving confidence interval for the mean or median.
These are a convenience wrappers around the Hmisc functions.
}
