% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{geom_confint}
\alias{geom_confint}
\title{Plot group confidence intervals}
\usage{
geom_confint(...)
}
\arguments{
\item{...}{Parameters passed to the geometry.}
}
\description{
Plot group confidence intervals
}
