% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_data}
\alias{graph_data}
\title{Add network metadata to edge and node lists}
\usage{
graph_data(nodes, edges, directed = FALSE)
}
\arguments{
\item{.data}{a data_frame containing the variables in the model}

\item{...}{variables to group by.}

\item{sort}{logical. Should the data be sorted on output.}
}
\description{
Add network metadata to edge and node lists
}
