% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{group_count}
\alias{group_count}
\title{Count number of elements in each group.}
\usage{
group_count(.data, ..., sort = FALSE)
}
\arguments{
\item{.data}{a data_frame containing the variables in the model}

\item{...}{variables to group by.}

\item{sort}{logical. Should the data be sorted on output.}
}
\description{
Count number of elements in each group.
}

