% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{group_summarize}
\alias{group_summarize}
\title{Summarize a data set by group or groups}
\usage{
group_summarize(.data, ...)
}
\arguments{
\item{.data}{a data_frame containing the variables in the model}

\item{...}{variables to group by.}
}
\description{
Summarize a data set by group or groups
}
