% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{lm_basic}
\alias{lm_basic}
\title{Fit a basic linear model}
\usage{
lm_basic(formula, data)
}
\arguments{
\item{formula}{an object of class \code{formula} (or one that can be
coerced to that class): a symbolic description of the
model to be fitted.}

\item{data}{a data_frame containing the variables in the model}
}
\description{
Fit a basic linear model
}
