% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{mplot}
\alias{mplot}
\title{A smarter map function}
\usage{
mplot(x, y, ..., data, maptype = "toner", zoom = 1, detail = 0,
  location = NULL)
}
\arguments{
\item{x}{variable name that defines the longitude}

\item{y}{variable name that defines the latitude}

\item{...}{other arguments passed to qplot}

\item{data}{the data frame containing longitude and latitude}

\item{maptype}{the type of map; either "toner" or "road"}

\item{zoom}{factor giving how much more area the map should cover than
the minimum area needed for the plot.}

\item{detail}{integer giving how much more detail should be given to the
map tiles compared to the default. Not applicable to road
maps. Negative integers are allowed, but are rarely useful.}

\item{location}{an option string describing the location to center the box on.}
}
\description{
A smarter map function
}
