% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{QuantileSummary}
\alias{QuantileSummary}
\alias{sm_quartiles}
\alias{sm_deciles}
\alias{sm_ventiles}
\alias{sm_percentiles}
\alias{sm_quantiles}
\title{Quantile summaries}
\usage{
sm_quartiles(x, name = NULL)

sm_deciles(x, name = NULL)

sm_ventiles(x, name = NULL)

sm_percentiles(x, name = NULL)

sm_quantiles(x, probs = seq(0, 1, 0.25), name = NULL)
}
\arguments{
\item{x}{the variable to summarize. Will be used deparsed to infer the
names of of the summaries.}

\item{name}{provide a prefix variable name for the output. Set to
\code{NULL} to infer from the input.}

\item{probs}{a complete list of probabilites to compute when calling the
\code{sm_quantiles} function.}
}
\description{
Summarize a variable through its quantiles.
}
