% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{reg_table}
\alias{reg_table}
\title{Compute regression table}
\usage{
reg_table(object, level = NULL, mu = 0)
}
\arguments{
\item{object}{a model object}

\item{level}{the confidence level}

\item{mu}{null hypothesis mean}
}
\description{
Compute regression table
}
