% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{sm_check_download}
\alias{sm_check_download}
\title{Check if a local file exists and otherwise download it}
\usage{
sm_check_download(src, dest)
}
\arguments{
\item{src}{URL where the file should be downloaded if it does not exist}

\item{dest}{local destination where the file should be stored}
}
\description{
Check if a local file exists and otherwise download it
}
