% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{sm_dunn_ll}
\alias{sm_dunn_ll}
\title{Compute Dunn Likelihood scores}
\usage{
sm_dunn_ll(object, group_name, token_name = "lemma")
}
\arguments{
\item{object}{data frame}

\item{group_name}{name of the variable containing the groups}

\item{token_name}{name of the variable containing the tokens}
}
\value{
a data frame
}
\description{
Compute Dunn Likelihood scores
}
