% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{sm_graph_layout}
\alias{sm_graph_layout}
\title{Add network metadata and layout information}
\usage{
sm_graph_layout(edges, node_name = "id")
}
\arguments{
\item{edges}{data frame describing the edges}

\item{node_name}{name of the node id column; defaults to "id"}
}
\description{
Add network metadata and layout information
}
