% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{sm_graph_metadata}
\alias{sm_graph_metadata}
\title{Add network metadata to edge and node lists}
\usage{
sm_graph_metadata(edges, node_name = "id")
}
\arguments{
\item{edges}{data frame describing the edges}

\item{node_name}{name of the node id column; defaults to "id"}
}
\description{
Add network metadata to edge and node lists
}
