% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{sm_kmeans}
\alias{sm_kmeans}
\title{Create a data frame of Kmeans clusters}
\usage{
sm_kmeans(df, clusters = 3, item_name = "document")
}
\arguments{
\item{df}{A data frame with a all numeric columns except the
one containing the document ID}

\item{clusters}{Integer. Number of clusters to identify.}

\item{item_name}{Name of the column containing the item id.}
}
\value{
A data frame with a column indicating the cluster ID.
}
\description{
Create a data frame of Kmeans clusters
}
