% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{sm_kwic}
\alias{sm_kwic}
\title{Keywords in Context}
\usage{
sm_kwic(term, text, n = Inf, ignore_case = TRUE, width = 20L)
}
\arguments{
\item{term}{search term as a string}

\item{text}{vector of text to search}

\item{n}{number of results to return; default is Inf}

\item{ignore_case}{should search ignore case? default is TRUE}

\item{width}{how many characters to show as context}
}
\value{
a character vector of the search results
}
\description{
Keywords in Context
}
