% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{sm_lda_topics}
\alias{sm_lda_topics}
\title{Compute Tidy LDA Topic Model}
\usage{
sm_lda_topics(X, num_topics = 16)
}
\arguments{
\item{X}{A term-frequency matrix}

\item{num_topics}{Integer. Number of topics to learn}
}
\value{
A named list of two data frames.
}
\description{
Compute Tidy LDA Topic Model
}
