% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{sm_ngram}
\alias{sm_ngram}
\title{N-Grams}
\usage{
sm_ngram(object, n_min = 1, n = 3, doc_var = "doc_id", token_var = "lemma")
}
\arguments{
\item{object}{data frame}

\item{n_min}{minimum number of terms}

\item{n}{maximum number of terms}

\item{doc_var}{name of the variable containing the document id}

\item{token_var}{name of the variable containing the tokens}
}
\value{
a data frame
}
\description{
N-Grams
}
