% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{sm_skip_ngram}
\alias{sm_skip_ngram}
\title{Skip N-Grams}
\usage{
sm_skip_ngram(
  object,
  n_min = 1,
  n = 3,
  k = 1,
  doc_var = "doc_id",
  token_var = "lemma"
)
}
\arguments{
\item{object}{data frame}

\item{n_min}{minimum number of terms}

\item{n}{maximum number of terms}

\item{k}{number of skips to consider}

\item{doc_var}{name of the variable containing the document id}

\item{token_var}{name of the variable containing the tokens}
}
\value{
a data frame
}
\description{
Skip N-Grams
}
