% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{sm_spectral_cluster}
\alias{sm_spectral_cluster}
\title{Create a data frame of Kmeans clusters}
\usage{
sm_spectral_cluster(X, max_depth = 4, balance = FALSE)
}
\arguments{
\item{X}{A TF-IDF matrix}

\item{max_depth}{Depth of the hierarchical clustering algorithm.
The maximum number of clusters will be 2^(max_depth).}

\item{balance}{Should clusters be made to have uniform size.}
}
\value{
A data frame.
}
\description{
Create a data frame of Kmeans clusters
}
