% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{sm_tidy_distance}
\alias{sm_tidy_distance}
\title{Create a data frame of pairwise (Euclidiean) distances}
\usage{
sm_tidy_distance(x, item_name = "document")
}
\arguments{
\item{x}{the matrix to compute distances between rows.}

\item{item_name}{name to give to the set of rows; defaults to "document"}
}
\description{
Create a data frame of pairwise (Euclidiean) distances
}
