% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{sm_tidy_glmnet}
\alias{sm_tidy_glmnet}
\title{Create a data frame of elastic net coefficents}
\usage{
sm_tidy_glmnet(X, y, item_name = "var", binomial = FALSE)
}
\arguments{
\item{X}{the data matrix, possibly sparse}

\item{y}{the response vector}

\item{item_name}{name to give to the set of rows; defaults to "var"}

\item{binomial}{logical; is this a binomial regression?}
}
\description{
Create a data frame of elastic net coefficents
}
