% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{sm_tidy_pca}
\alias{sm_tidy_pca}
\title{Create a data frame of principal components}
\usage{
sm_tidy_pca(x, n = 3, scale = TRUE, item_name = "document")
}
\arguments{
\item{x}{the matrix to compute principal components from.}

\item{n}{number of components to compute}

\item{scale}{logical; should columns by scaled before PCA}

\item{item_name}{name to give to the set of rows; defaults to "document"}
}
\description{
Create a data frame of principal components
}
