% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{sm_tidy_umap}
\alias{sm_tidy_umap}
\title{Create a data frame of umap components}
\usage{
sm_tidy_umap(x, n = 2, random_state = 1, item_name = "document", ...)
}
\arguments{
\item{x}{the matrix to compute principal components from.}

\item{n}{number of components to compute}

\item{random_state}{integer to set the random state of the algorithm}

\item{item_name}{name to give to the set of rows; defaults to "document"}

\item{...}{other options passed to the umap function}
}
\description{
Create a data frame of umap components
}
