% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{sm_to_pixel}
\alias{sm_to_pixel}
\title{Unravel Nested Dataset of Images into Pixels}
\usage{
sm_to_pixel(object, img_col_name = "img")
}
\arguments{
\item{object}{the input data frame}

\item{img_col_name}{name of the column containing the image arrays}
}
\value{
a tibble in wide format with term frequencies and tf-idf values.
}
\description{
Unravel Nested Dataset of Images into Pixels
}
