% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{term_df_to_matrix}
\alias{term_df_to_matrix}
\title{Convert a data frame of terms to a sparse matrix}
\usage{
term_df_to_matrix(x, min_df = 0, max_df = 1, max_features = 10000,
  scale = FALSE, vocabulary = NULL)
}
\arguments{
\item{x}{data frame}

\item{min_df}{minimum fraction of documents to cut off at}

\item{max_df}{maximum fraction of documents to cut off at}

\item{max_features}{maximum number of features}

\item{scale}{should data be scaled by rows}

\item{vocabulary}{optional vocabulary vector; if given
min_df, max_df, and max_features are ignored}
}
\description{
Convert a data frame of terms to a sparse matrix
}
