% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{term_list_to_df}
\alias{term_list_to_df}
\title{Convert a list of terms to a data frame}
\usage{
term_list_to_df(tokens)
}
\arguments{
\item{tokens}{list of tokens}
}
\description{
Convert a list of terms to a data frame
}
