% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_sm}
\alias{theme_sm}
\title{Minimal theme perfect for presentations}
\usage{
theme_sm()
}
\description{
Return a minimalistic theme based on Tufte's
*Visual Display of Quantitative Information*, with some custom tweaks
}
