% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_data}
\alias{graph_data}
\title{Add network metadata to edge and node lists}
\usage{
graph_data(edges, nodes = NULL, nodes2 = NULL, directed = FALSE,
  layout = NULL)
}
\arguments{
\item{edges}{data frame describing the edges}

\item{nodes}{data frame describing the nodes}

\item{directed}{logical. Is this a directed graph?}

\item{bipartite}{logical. Is this a directed graph?}
}
\description{
Add network metadata to edge and node lists
}
