% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{SingleSummary}
\alias{SingleSummary}
\alias{sm_mean}
\alias{sm_median}
\alias{sm_sd}
\alias{sm_sum}
\alias{sm_min}
\alias{sm_max}
\alias{sm_cor}
\alias{sm_cov}
\alias{sm_var}
\alias{sm_na_count}
\alias{sm_count}
\alias{sm_paste}
\title{Numeric Summaries}
\usage{
sm_mean(x, name = NULL)

sm_median(x, name = NULL)

sm_sd(x, name = NULL)

sm_sum(x, name = NULL)

sm_min(x, name = NULL)

sm_max(x, name = NULL)

sm_cor(x, y, name = NULL)

sm_cov(x, y, name = NULL)

sm_var(x, y, name = NULL)

sm_na_count(x, name = NULL)

sm_count(x, name = NULL)

sm_paste(x, name = NULL, sep = "; ")
}
\arguments{
\item{x}{the variable to summarize. Will be used deparsed to infer the
names of of the summaries.}

\item{name}{provide a prefix variable name for the output. Set to
\code{NULL} to infer from the input.}

\item{y}{secomd variable to summarize, for correlation, covariance, and
variance.}

\item{sep}{string to use to collapse text values with `sm_paste`}
}
\description{
Summarize a variable through a single numeric summary, such as the mean
standard deviation, and sum.
}
